Move serialization utils
