+++ language-support/jobrunr-kotlin-21-support/src/main/kotlin/org/jobrunr/kotlin/serialization/JobDashboardProgressSerializer.kt
@@ -5,7 +5,7 @@ import kotlinx.serialization.builtins.serializer
 import kotlinx.serialization.descriptors.buildClassSerialDescriptor
 import kotlinx.serialization.encoding.*
 import org.jobrunr.jobs.context.JobDashboardProgressBar
-import org.jobrunr.kotlin.utils.mapper.ContextualFallbackSerializer
+import org.jobrunr.kotlin.serialization.utils.ContextualFallbackSerializer
 
 object JobDashboardProgressSerializer : KSerializer<JobDashboardProgressBar.JobDashboardProgress>, ContextualFallbackSerializer.PolymorphicContinuationSerializer {
 	override val descriptor = buildClassSerialDescriptor(JobDashboardProgressBar.JobDashboardProgress::class.qualifiedName!!) {
