+++ language-support/jobrunr-kotlin-21-support/src/main/kotlin/org/jobrunr/kotlin/serialization/utils/AnySerializer.kt
@@ -1,4 +1,4 @@
-package org.jobrunr.kotlin.utils.mapper
+package org.jobrunr.kotlin.serialization.utils
 
 import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.KSerializer
@@ -60,4 +60,3 @@ open class AnySerializer<T : Any> : KSerializer<Any> {
 		error("Unexpected end of input")
 	}
 }
-
